+++ codingame/utils.py
@@ -1,6 +1,6 @@
 import re
 import typing
-from datetime import datetime
+from datetime import datetime, timezone
 
 from .exceptions import LoginRequired
 
@@ -103,8 +103,8 @@ DT_FORMAT_2 = "%b %d, %Y, %I:%M:%S %p"  # see issue #23
 
 def to_datetime(data: typing.Optional[typing.Union[int, str]]) -> datetime:
     if isinstance(data, int):
-        return datetime.utcfromtimestamp(data / 1000.0)
-    elif isinstance(data, str):
+        return datetime.fromtimestamp(data / 1000.0, timezone.utc)
+    elif isinstance(data, str):  # pragma: no cover
         try:
             return datetime.strptime(data, DT_FORMAT_1)
         except ValueError: