+++ scripts/format.sh
@@ -6,11 +6,11 @@ folders="codingame tests examples docs setup.py"
 set -x
 
 # put every import on one line for autoflake remove unused imports
-isort $folders --force-single-line-imports
+python3 -m isort $folders --force-single-line-imports
 # remove unused imports and variables
-autoflake $folders --remove-all-unused-imports --recursive --remove-unused-variables --in-place --exclude=__init__.py
+python3 -m autoflake $folders --remove-all-unused-imports --recursive --remove-unused-variables --in-place --exclude=__init__.py
 # resort imports
-isort $folders
+python3 -m isort $folders
 
 # format code
-black $folders --line-length 80
+python3 -m black $folders --line-length 80