+++ tests/sync/test_client.py
@@ -358,7 +358,7 @@ def test_client_mark_notifications_as_seen(auth_client: SyncClient, mock_http):
     mock_http(
         auth_client._state.http,
         "mark_notifications_as_seen",
-        int(datetime.datetime.utcnow().timestamp() * 1000),
+        int(datetime.datetime.now(datetime.timezone.utc).timestamp() * 1000),
     )
     seen_date = auth_client.mark_notifications_as_seen(
         [notification, notification.id]
@@ -367,7 +367,7 @@ def test_client_mark_notifications_as_seen(auth_client: SyncClient, mock_http):
     assert notification.seen
     assert notification.seen_date == seen_date
     assert notification.seen_date.timestamp() == pytest.approx(
-        datetime.datetime.utcnow().timestamp(), abs=10_000
+        datetime.datetime.now(datetime.timezone.utc).timestamp(), abs=10_000
     )  # 10 seconds should be enough